requirements.txt
네, 이 코드에서 필요한 외부 패키지를 기준으로 requirements.txt를 만들면 아래처럼 됩니다
python-dotenv
openai
requests
pymongo